+++ Mint/ParasitismShooting/Character/Shot/ShotFactory.h
@@ -38,6 +38,6 @@ public:
 	void ShotDelete();
 	void OutofScreen(void);
 
-	double SetAngle(std::string shotType, Vector2f pos, int shooter,int cnt,int);
+	float SetAngle(std::string shotType, Vector2f pos, int shooter,int cnt,int);
 };
