+++ ParasitismShooting/Character/Shot/Shot.h
@@ -34,7 +34,7 @@ enum SHOOTER
 struct ShotST
 {
 	Vector2f cpos;
-	double angle;
+	float angle;
 	int speed;
 	int movePtn;
 	int level;
@@ -57,7 +57,7 @@ protected:
 
 	ShotST shotst;
 
-	double SetTracking(std::string shotType, Vector2f pos, int shooter, double oldangle);
+	float SetTracking(std::string shotType, Vector2f pos, int shooter, float oldangle);
 
 public:
 	virtual ~Shot();
